requirements.txt
asgiref
Django
django-chunked-upload
django-ratelimit
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
PyMySQL
pytz
sorl-thumbnail
sqlparse
PyJWT
flask
pandas
numpy
requests
python-docx
openai